+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/ide/colors/HuffColors.kt
@@ -16,10 +16,45 @@ enum class HuffColor(
     HuffBundle.messagePointer("settings.huff.color.block_comment"),
     Default.BLOCK_COMMENT,
   ),
+  STRING(HuffBundle.messagePointer("settings.huff.color.string"), Default.STRING),
   HEXCODE(HuffBundle.messagePointer("settings.huff.color.hexcode"), Default.NUMBER),
   DEFINE(HuffBundle.messagePointer("settings.huff.color.define"), Default.KEYWORD),
   INCLUDE(HuffBundle.messagePointer("settings.huff.color.include"), Default.CLASS_REFERENCE),
-  BRACES(HuffBundle.messagePointer("settings.huff.color.braces"), Default.BRACES);
+  BRACES(HuffBundle.messagePointer("settings.huff.color.braces"), Default.BRACES),
+  BUILDIN_FN(HuffBundle.messagePointer("settings.huff.color.buildin_fn"), Default.STATIC_METHOD),
+  ARITHMETIC_AND_LOGICAL(
+    HuffBundle.messagePointer("settings.huff.color.arithmetic_and_logical"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  STACK_MANAGEMENT(
+    HuffBundle.messagePointer("settings.huff.color.stack_management"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  ENVIRONMENTAL_INFORMATION(
+    HuffBundle.messagePointer("settings.huff.color.environmental_information"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  STORAGE_AND_MEMORY_ACCESS(
+    HuffBundle.messagePointer("settings.huff.color.storage_and_memory_access"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  FLOW_CONTROL(
+    HuffBundle.messagePointer("settings.huff.color.flow_control"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  SYSTEM_AND_CALL(
+    HuffBundle.messagePointer("settings.huff.color.system_and_call"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  LOGGING(HuffBundle.messagePointer("settings.huff.color.logging"), Default.HIGHLIGHTED_REFERENCE),
+  MISCELLANEOUS(
+    HuffBundle.messagePointer("settings.huff.color.miscellaneous"),
+    Default.HIGHLIGHTED_REFERENCE,
+  ),
+  GAS_AND_COST_MANAGEMENT(
+    HuffBundle.messagePointer("settings.huff.color.gas_and_cost_management"),
+    Default.HIGHLIGHTED_REFERENCE,
+  );
 
   val textAttributesKey = TextAttributesKey.createTextAttributesKey("run.huff.$name", default)
   val attributesDescriptor = AttributesDescriptor(humanName, textAttributesKey)