+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/OpcodeNamePsiNode.kt
@@ -1,0 +1,6 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi
+
+import com.intellij.lang.ASTNode
+import org.antlr.intellij.adaptor.psi.ANTLRPsiNode
+
+open class OpcodeNamePsiNode(node: ASTNode) : ANTLRPsiNode(node)